refactor: ChartCanvas separate types & ctx
