+++ packages/core/src/ChartCanvas/ChartCanvas.tsx
@@ -1,19 +1,19 @@
 import { extent as d3Extent, max, min } from "d3-array";
 import { ScaleContinuousNumeric, ScaleTime } from "d3-scale";
 import * as React from "react";
-import { clearCanvas, functor, head, identity, isDefined, isNotDefined, last, shallowEqual } from "./utils";
-import { IZoomAnchorOptions, mouseBasedZoomAnchor } from "./zoom";
+import { clearCanvas, functor, head, identity, isDefined, isNotDefined, last, shallowEqual } from "../utils";
+import { mouseBasedZoomAnchor } from "../zoom";
 import {
-    ChartConfig,
     getChartConfigWithUpdatedYScales,
     getCurrentCharts,
     getCurrentItem,
     getNewChartConfig,
-} from "./utils/ChartDataUtil";
-import { EventCapture } from "./EventCapture";
-import { CanvasContainer, ICanvasContexts } from "./CanvasContainer";
-import evaluator from "./utils/evaluator";
-import type { MoreProps } from "./MoreProps";
+} from "../utils/ChartDataUtil";
+import { EventCapture } from "../EventCapture";
+import { CanvasContainer } from "../CanvasContainer";
+import evaluator from "../utils/evaluator";
+import { ChartCanvasContext } from '.'
+import type { ChartCanvasContextType, ChartCanvasProps, ChartCanvasState, MutableState, Subscription } from '.';
 
 const CANDIDATES_FOR_RESET = ["seriesName"];
 
@@ -67,56 +67,6 @@ const getCursorStyle = () => {
 	}`;
     return <style type="text/css">{tooltipStyle}</style>;
 };
-
-export interface ChartCanvasContextType<TXAxis extends number | Date> {
-    width: number;
-    height: number;
-    margin: { top: number; right: number; bottom: number; left: number };
-    chartId: number | string;
-    getCanvasContexts?: () => ICanvasContexts | undefined;
-    xScale: Function;
-    ratio: number;
-    // Not sure if it should be optional
-    xAccessor: (data: any) => TXAxis;
-    displayXAccessor: (data: any) => TXAxis;
-    xAxisZoom?: (newDomain: any) => void;
-    yAxisZoom?: (chartId: string, newDomain: any) => void;
-    redraw: () => void;
-    plotData: any[];
-    fullData: any[];
-    chartConfigs: ChartConfig[];
-    morePropsDecorator?: () => void;
-    generateSubscriptionId?: () => number;
-    getMutableState: () => {};
-    amIOnTop: (id: string | number) => boolean;
-    subscribe: (id: string | number, rest: any) => void;
-    unsubscribe: (id: string | number) => void;
-    setCursorClass: (className: string | null | undefined) => void;
-}
-
-// eslint-disable-next-line @typescript-eslint/no-empty-function
-const noop = () => {};
-export const chartCanvasContextDefaultValue: ChartCanvasContextType<number | Date> = {
-    amIOnTop: () => false,
-    chartConfigs: [],
-    chartId: 0,
-    ratio: 0,
-    displayXAccessor: () => 0,
-    fullData: [],
-    getMutableState: () => ({}),
-    height: 0,
-    margin: { top: 0, right: 0, bottom: 0, left: 0 },
-    plotData: [],
-    setCursorClass: noop,
-    subscribe: noop,
-    unsubscribe: noop,
-    redraw: noop,
-    width: 0,
-    xAccessor: () => 0,
-    xScale: noop,
-};
-export const ChartCanvasContext =
-    React.createContext<ChartCanvasContextType<number | Date>>(chartCanvasContextDefaultValue);
 
 const getDimensions = <TXAxis extends number | Date>(props: ChartCanvasProps<TXAxis>) => {
     const { margin, height, width } = props;
@@ -329,97 +279,6 @@ const isInteractionEnabled = (
     const interaction = !isNaN(xScale(xAccessor(head(data)))) && isDefined(xScale.invert);
     return interaction;
 };
-
-export interface ChartCanvasProps<TXAxis extends number | Date> {
-    readonly clamp?:
-        | boolean
-        | ("left" | "right" | "both")
-        | ((domain: [number, number], items: [number, number]) => [number, number]);
-    readonly className?: string;
-    readonly children?: React.ReactNode;
-    readonly data: any[];
-    readonly defaultFocus?: boolean;
-    readonly disableInteraction?: boolean;
-    readonly disablePan?: boolean;
-    readonly disableZoom?: boolean;
-    readonly displayXAccessor?: (data: any) => TXAxis;
-    readonly flipXScale?: boolean;
-    readonly height: number;
-    readonly margin: {
-        bottom: number;
-        left: number;
-        right: number;
-        top: number;
-    };
-    readonly maintainPointsPerPixelOnResize?: boolean;
-    readonly minPointsPerPxThreshold?: number;
-    readonly mouseMoveEvent?: boolean;
-    /**
-     * Called when panning left past the first data point.
-     */
-    readonly onLoadAfter?: (start: TXAxis, end: TXAxis) => void;
-    /**
-     * Called when panning right past the last data point.
-     */
-    readonly onLoadBefore?: (start: TXAxis, end: TXAxis) => void;
-    /**
-     * Click event handler.
-     */
-    readonly onClick?: React.MouseEventHandler<HTMLDivElement>;
-    /**
-     * Double click event handler.
-     */
-    readonly onDoubleClick?: React.MouseEventHandler<HTMLDivElement>;
-    readonly padding?:
-        | number
-        | {
-              bottom: number;
-              left: number;
-              right: number;
-              top: number;
-          };
-    readonly plotFull?: boolean;
-    readonly pointsPerPxThreshold?: number;
-    readonly postCalculator?: (plotData: any[]) => any[];
-    readonly ratio: number;
-    readonly seriesName: string;
-    readonly useCrossHairStyleCursor?: boolean;
-    readonly width: number;
-    readonly xAccessor: (data: any) => TXAxis;
-    readonly xExtents: ((data: any[]) => [TXAxis, TXAxis]) | (((data: any[]) => TXAxis) | TXAxis)[];
-    readonly xScale: ScaleContinuousNumeric<number, number> | ScaleTime<number, number>;
-    readonly zIndex?: number;
-    readonly zoomAnchor?: (options: IZoomAnchorOptions<any, TXAxis>) => TXAxis;
-    readonly zoomMultiplier?: number;
-}
-
-interface ChartCanvasState<TXAxis extends number | Date> {
-    lastProps?: ChartCanvasProps<TXAxis>;
-    propIteration?: number;
-    xAccessor: (data: any) => TXAxis;
-    displayXAccessor?: any;
-    filterData?: any;
-    chartConfigs: ChartConfig[];
-    plotData: any[];
-    xScale: ScaleContinuousNumeric<number, number> | ScaleTime<number, number>;
-    fullData: any[];
-}
-
-interface Subscription {
-    id: string;
-    getPanConditions: () => {
-        draggable: boolean;
-        panEnabled: boolean;
-    };
-    draw: (props: { trigger: string } | { force: boolean }) => void;
-    listener: (type: string, newMoreProps: MoreProps | undefined, state: any, e: any) => void;
-}
-
-interface MutableState {
-    mouseXY: [number, number];
-    currentItem: any;
-    currentCharts: string[];
-}
 
 export class ChartCanvas<TXAxis extends number | Date> extends React.Component<
     ChartCanvasProps<TXAxis>,
